requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
numpy
opencv-python-headless
python-multipart